Renaming and relocating org-related UI components
