+++ src/dispatch/static/dispatch/src/router/config.js
@@ -280,7 +280,7 @@ export const protectedRoute = [
           name: "OrganizationMemberTable",
           meta: { title: "Members", subMenu: "organization", group: "organization" },
           component: () =>
-            import(/* webpackChunkName: "users-table" */ "@/auth/OrganizationMemberTable.vue"),
+            import(/* webpackChunkName: "users-table" */ "@/organization/OrganizationMemberTable.vue"),
         },
         {
           path: "projects",
